+++ CHANGELOG.md
@@ -421,3 +421,4 @@
 - 2024-06-03: Add qualify, email opener, summaries, send, +3 db
 - 2024-06-04: Pagination sender, mod, leads, embeddings, +6
 - 2024-06-05: Add integration, user, bettercontact, mutate queries, +3 types
+- 2024-06-06: Valid qualify, leads, send, email opener, +4